# Conflicts:
#	app/build.gradle.kts
#	app/src/main/AndroidManifest.xml

+++ app/src/main/java/org/thunderdog/challegram/component/attach/MediaBottomGalleryController.java
@@ -790,7 +790,7 @@ public class MediaBottomGalleryController extends MediaBottomBaseController<Medi
   }
 
   private String awaitingQuery;
-  private int bingUserId;
+  private long bingUserId;
   private boolean bingUserLoading;
 
   private void searchInternal (final String q) {
